Simplify interface for exportRows
