+++ packages/components/src/index.ts
@@ -121,7 +121,7 @@ import {
     getSelected,
     getSelectedData,
     getSelection,
-    getGridExportParams,
+    getQueryModelExportParams,
     gridExport,
     gridIdInvalidate,
     gridInit,
@@ -400,7 +400,7 @@ export {
     getSelected,
     getSelectedData,
     getSelection,
-    getGridExportParams,
+    getQueryModelExportParams,
     gridExport,
     gridInit,
     gridShowError,
